Added subsetting code to example_utils.py.

+++ ml4tc/scripts/create_example_files.py
@@ -12,6 +12,8 @@ from ml4tc.utils import general_utils
 SATELLITE_DIR_ARG_NAME = 'input_satellite_dir_name'
 SHIPS_DIR_ARG_NAME = 'input_ships_dir_name'
 YEAR_ARG_NAME = 'year'
+TIME_INTERVAL_ARG_NAME = 'satellite_time_interval_sec'
+COMPRESS_ARG_NAME = 'compress_output_files'
 OUTPUT_DIR_ARG_NAME = 'output_dir_name'
 
 SATELLITE_DIR_HELP_STRING = (
@@ -23,6 +25,8 @@ SHIPS_DIR_HELP_STRING = (
     'by `ships_io.find_file` and read by `ships_io.read_file`.'
 )
 YEAR_HELP_STRING = 'Example files will be created only for this year.'
+TIME_INTERVAL_HELP_STRING = 'Time interval for satellite data (seconds).'
+COMPRESS_HELP_STRING = 'Boolean flag.  If 1 (0), will (not) gzip output files.'
 OUTPUT_DIR_HELP_STRING = (
     'Name of top-level directory for learning examples.  Files will be written '
     'here by `example_io.write_file`, with exact names determined by '
@@ -42,12 +46,22 @@ INPUT_ARG_PARSER.add_argument(
     '--' + YEAR_ARG_NAME, type=int, required=True, help=YEAR_HELP_STRING
 )
 INPUT_ARG_PARSER.add_argument(
+    '--' + TIME_INTERVAL_ARG_NAME, type=int, required=True,
+    help=TIME_INTERVAL_HELP_STRING
+)
+INPUT_ARG_PARSER.add_argument(
+    '--' + COMPRESS_ARG_NAME, type=int, required=True,
+    help=COMPRESS_HELP_STRING
+)
+INPUT_ARG_PARSER.add_argument(
     '--' + OUTPUT_DIR_ARG_NAME, type=str, required=True,
     help=OUTPUT_DIR_HELP_STRING
 )
 
 
-def _run(top_satellite_dir_name, top_ships_dir_name, year, top_output_dir_name):
+def _run(top_satellite_dir_name, top_ships_dir_name, year,
+         satellite_time_interval_sec, compress_output_files,
+         top_output_dir_name):
     """Creates example files by merging satellite and SHIPS files.
 
     This is effectively the main method.
@@ -55,6 +69,8 @@ def _run(top_satellite_dir_name, top_ships_dir_name, year, top_output_dir_name):
     :param top_satellite_dir_name: See documentation at top of file.
     :param top_ships_dir_name: Same.
     :param year: Same.
+    :param satellite_time_interval_sec: Same.
+    :param compress_output_files: Same.
     :param top_output_dir_name: Same.
     """
 
@@ -110,6 +126,10 @@ def _run(top_satellite_dir_name, top_ships_dir_name, year, top_output_dir_name):
             satellite_table_xarray=this_satellite_table_xarray,
             ships_table_xarray=this_ships_table_xarray
         )
+        this_example_table_xarray = example_utils.subset_satellite_times(
+            example_table_xarray=this_example_table_xarray,
+            time_interval_sec=satellite_time_interval_sec
+        )
 
         print('Writing data to: "{0:s}"...\n'.format(this_example_file_name))
         example_io.write_file(
@@ -117,8 +137,9 @@ def _run(top_satellite_dir_name, top_ships_dir_name, year, top_output_dir_name):
             netcdf_file_name=this_example_file_name
         )
 
-        general_utils.compress_file(this_example_file_name)
-        os.remove(this_example_file_name)
+        if compress_output_files:
+            general_utils.compress_file(this_example_file_name)
+            os.remove(this_example_file_name)
 
 
 if __name__ == '__main__':
@@ -130,5 +151,11 @@ if __name__ == '__main__':
         ),
         top_ships_dir_name=getattr(INPUT_ARG_OBJECT, SHIPS_DIR_ARG_NAME),
         year=getattr(INPUT_ARG_OBJECT, YEAR_ARG_NAME),
+        satellite_time_interval_sec=getattr(
+            INPUT_ARG_OBJECT, TIME_INTERVAL_ARG_NAME
+        ),
+        compress_output_files=bool(
+            getattr(INPUT_ARG_OBJECT, COMPRESS_ARG_NAME)
+        ),
         top_output_dir_name=getattr(INPUT_ARG_OBJECT, OUTPUT_DIR_ARG_NAME)
     )
